backend/requirements.txt
fastapi
uvicorn
fastapi-cors
transformers
torch
Pillow
einops
